# apps/backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Staylark API",
        "version": "1.0.0",
        "description": "Interactive API reference for the Staylark backend.\n\n## Authentication\nMost endpoints require a Bearer token. To obtain one:\n\n```bash\ncurl -X POST '<BASE_URL>/api/auth/sign-in/email' \\\n  -H 'Content-Type: application/json' \\\n  -d '{\"email\":\"you@example.com\",\"password\":\"yourpassword\"}'\n```\n\nCopy the `token` from the response, click **Authorize** above, and paste it."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "session-token",
                "description": "Better Auth session token. Obtain via `POST /api/auth/sign-in/email`, then paste the returned `token` here."
            }
        },
        "schemas": {
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "number"
                    },
                    "errorCode": {
                        "type": "string"
                    },
                    "details": {}
                },
                "required": ["error", "message", "statusCode"],
                "example": {
                    "error": "Validation Error",
                    "message": "guests: Number must be greater than 0",
                    "statusCode": 400
                }
            },
            "Property": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": ["APARTMENT", "HOUSE", "HOTEL"]
                    },
                    "city": {
                        "type": "string",
                        "minLength": 1
                    },
                    "address": {
                        "type": "string",
                        "minLength": 1
                    },
                    "nightlyPrice": {
                        "type": "number",
                        "exclusiveMinimum": 0
                    },
                    "minNightlyPrice": {
                        "type": ["number", "null"],
                        "exclusiveMinimum": 0
                    },
                    "maxNightlyPrice": {
                        "type": ["number", "null"],
                        "exclusiveMinimum": 0
                    },
                    "maxGuests": {
                        "type": "integer",
                        "exclusiveMinimum": 0
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "photos": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        }
                    },
                    "managerId": {
                        "type": ["string", "null"],
                        "minLength": 1
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "description",
                    "type",
                    "city",
                    "address",
                    "nightlyPrice",
                    "maxGuests",
                    "amenities",
                    "photos",
                    "createdAt",
                    "updatedAt"
                ]
            },
...
        },
...
    },
    "paths": {
        "/api/properties/amenities": {
            "get": {
                "tags": ["Properties"],
                "summary": "List all distinct amenities",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Amenity names",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/cities": {
            "get": {
                "tags": ["Properties"],
                "summary": "List all distinct cities",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City names",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties": {
            "get": {
                "tags": ["Properties"],
                "summary": "List properties (filterable, paginated)",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": false,
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string",
                            "enum": ["APARTMENT", "HOUSE", "HOTEL"]
                        },
                        "required": false,
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "number",
                            "exclusiveMinimum": 0
                        },
                        "required": false,
                        "name": "minPrice",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "number",
                            "exclusiveMinimum": 0
                        },
                        "required": false,
                        "name": "maxPrice",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "required": false,
                        "name": "amenities",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string",
                            "enum": ["title", "city", "basePrice", "createdAt"]
                        },
                        "required": false,
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "string",
                            "enum": ["asc", "desc"]
                        },
                        "required": false,
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "default": 1
                        },
                        "required": false,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "schema": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "maximum": 100,
                            "default": 20
                        },
                        "required": false,
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        },
                                        "total": {
                                            "type": "number"
                                        },
                                        "page": {
                                            "type": "number"
                                        },
                                        "limit": {
                                            "type": "number"
                                        }
                                    },
                                    "required": ["data", "total", "page", "limit"]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Properties"],
                "summary": "Create a property (ADMIN only)",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProperty"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created property",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/{id}": {
            "get": {
                "tags": ["Properties"],
                "summary": "Get a property by ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "description": "Property ID"
                        },
                        "required": true,
                        "description": "Property ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": ["Properties"],
                "summary": "Update a property (ADMIN only)",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "description": "Property ID"
                        },
                        "required": true,
                        "description": "Property ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProperty"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated property",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Property"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Properties"],
                "summary": "Soft-delete a property (ADMIN only)",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "description": "Property ID"
                        },
                        "required": true,
                        "description": "Property ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted (no content)"
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/{id}/availability": {
            "get": {
                "tags": ["Properties"],
                "summary": "Get a property's booked date ranges",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string",
                            "description": "Property ID"
                        },
                        "required": true,
                        "description": "Property ID",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booked date ranges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "checkIn": {
                                                "type": "string"
                                            },
                                            "checkOut": {
                                                "type": "string"
                                            }
                                        },
                                        "required": ["checkIn", "checkOut"]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/import/template": {
            "get": {
                "tags": ["Properties"],
                "summary": "Download the bulk-import spreadsheet template (MANAGER only)",
                "description": "Generated per request so the amenity suggestions reflect the live catalogue. Contains three sheets: Properties (fill this in), Reference and Example.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "XLSX template",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/properties/import": {
            "post": {
                "tags": ["Properties"],
                "summary": "Bulk-create properties from a spreadsheet (MANAGER only)",
                "description": "All-or-nothing: any invalid row rejects the whole file and nothing is written. Every created property is assigned to the uploading manager; a managerId column is not accepted. Photo URLs are validated for reachability during the request, then re-hosted on Cloudinary in the background.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": ["file"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Properties created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportPropertiesResult"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The file was rejected. No properties were created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "enum": [422]
                                        },
                                        "errorCode": {
                                            "type": "string",
                                            "enum": ["IMPORT_VALIDATION_FAILED"]
                                        },
                                        "details": {
                                            "type": "object",
                                            "properties": {
                                                "rowCount": {
                                                    "type": "number"
                                                },
                                                "errors": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ImportRowError"
                                                    }
                                                }
                                            },
                                            "required": ["rowCount", "errors"]
                                        }
                                    },
                                    "required": ["error", "message", "statusCode", "errorCode", "details"]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}